Added phantom switch mode.
Phantom switch is kind of a remote for a real switch. Upon pressing a
button, it only sends an MQTT command to the configured switch instead
of driving a real switch.
Phantom switches are not tracking any internal state so they are not
supposed to be added to OpenHAB panel and controlled remotely.

* Added means to configure a phantom switch,
* Added 'short tap'/'long tap' MQTT channel cmds,
* In phantom switch mode, OnShortTap()/OnLongTap() only send a tap cmd
  to a configured real switch.

Signed-off-by: Lukasz Lasek <lukasz.lasek@gmail.com>